+++ src/inspect_scout/_view/_transcripts_helpers.py
